requirements.txt
fastapi==0.104.1
pymongo[srv]==4.6.0
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
